Eliminated nearly every reference to PageManager with ContentManager in the code, plus or minus about two references. JSPs not checked yet.
git-svn-id: https://svn.apache.org/repos/asf/incubator/jspwiki/trunk@762298 13f79535-47bb-0310-9956-ffa450edef68

+++ tests/java/org/apache/wiki/plugin/PluginManagerTest.java
@@ -57,14 +57,16 @@ public class PluginManagerTest extends TestCase
         props.load( TestEngine.findTestProperties() );
 
         engine = new TestEngine(props);
-        engine.deletePage( "Testpage" );
         context = engine.getWikiContextFactory().newViewContext( engine.createPage( "Testpage" ) );
         manager = new PluginManager( engine, props );
     }
 
     public void tearDown() throws Exception
     {
-        engine.deletePage("Testpage");
+        if ( engine.pageExists( "Testpage" ) )
+        {
+            engine.deletePage("Testpage");
+        }
         engine.shutdown();
     }
